EduKids/requirements_ai.txt
# Requirements pour la fonctionnalité IA Avatar

# HuggingFace pour Stable Diffusion Inpainting
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
torch>=2.1.0
safetensors>=0.4.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# HTTP requests
requests>=2.31.0

# Existing requirements (à ajouter à requirements.txt principal)
# Django>=4.2.0
# djangorestframework>=3.14.0
# python-decouple>=3.8
